fix: build issue by removing API calls from prerendering process
