+++ pages/quests/[id].tsx
@@ -1,22 +1,14 @@
-import axios from 'axios';
 import SingleQuestPage from '@components/SingleQuestPage';
-import Quest from '@interfaces/Quest.interface';
 import { useQuery } from '@tanstack/react-query';
 import { GetStaticPaths, GetStaticProps } from 'next';
-import { fetchQuests } from 'pages';
 import { ISingleQuest } from '@interfaces/Pages.interface';
 import { ParsedUrlQuery } from 'querystring';
-
-const baseURL = 'http://localhost:3000';
-
-function fetchQuest(questId: string): Promise<Quest> {
-	return axios(`${baseURL}/api/quests/${questId}`).then((res) => res.data);
-}
+import { fetchQuestFromApi, getAllQuests, getQuestById } from 'src/services/data.service';
 
 function SingleQuest({ quest }: ISingleQuest) {
 	const { isLoading, error, data } = useQuery({
 		queryKey: ['quest', quest.id],
-		queryFn: () => fetchQuest(quest.id.toString()),
+		queryFn: () => fetchQuestFromApi(quest.id.toString()),
 		initialData: quest
 	});
 
@@ -27,7 +19,7 @@ function SingleQuest({ quest }: ISingleQuest) {
 }
 
 export const getStaticPaths: GetStaticPaths = async () => {
-	const quests = await fetchQuests();
+	const quests = await getAllQuests();
 
 	const paths = quests.map((quest) => ({
 		params: { id: quest.id.toString() }
@@ -42,7 +34,7 @@ interface IParams extends ParsedUrlQuery {
 
 export const getStaticProps: GetStaticProps<ISingleQuest, IParams> = async ({ params }) => {
 	const { id: questId } = params as IParams;
-	let result = await fetchQuest(questId);
+	let result = await getQuestById(questId);
 
 	return {
 		props: {
